opensearch-mcp-server/requirements.txt
opensearch-py>=2.0.0
python-dotenv>=1.0.0
mcp-server>=0.1.0
pydantic>=2.0.0
requests>=2.28.0 
strands-agents>=0.1.0
strands-agents-tools>=0.1.0